+++ src/main/java/com/zktr/crmproject/pojos/Pcrdetail.java
@@ -3,12 +3,13 @@ package com.zktr.crmproject.pojos;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 public class Pcrdetail {
     private int pcrdId;
     private Integer quantity;
-    private Integer uintPrice;
+    private BigDecimal uintPrice;
     private String remarks;
     @JsonIgnoreProperties("pcrdetail")
     private Purchasereturn purchasereturn;
@@ -39,11 +40,11 @@ public class Pcrdetail {
 
     @Basic
     @Column(name = "uint_price")
-    public Integer getUintPrice() {
+    public BigDecimal getUintPrice() {
         return uintPrice;
     }
 
-    public void setUintPrice(Integer uintPrice) {
+    public void setUintPrice(BigDecimal uintPrice) {
         this.uintPrice = uintPrice;
     }
